Move all redux related api calls to complex.actions.ts
